+++ test/cases/_preprocessor/js-tmpl-hbs-partials-in-hbs-js/src/app.js
@@ -1,0 +1,12 @@
+import tmpl from './app.hbs';
+
+const locals = {
+  helloName: 'World',
+  name: 'Max',
+  age: 21,
+  navText: 'Navigation',
+};
+
+document.getElementById('app').innerHTML = tmpl(locals);
+
+console.log('>> app');